Rename projecto to Slicker
Also rename XorT to EitherT in respect of coming scala 2.12 update
