+++ README.md
@@ -1,3 +1,3 @@
-# slick-repository [![Build Status](https://travis-ci.org/ImLiar/slick-repository.svg?branch=travis-ci)](https://travis-ci.org/ImLiar/slick-repository)
+# Slicker [![Build Status](https://travis-ci.org/ImLiar/slick-repository.svg?branch=travis-ci)](https://travis-ci.org/ImLiar/slick-repository)
 ---
 Putting Slick into repository
