--- slick-repository-postgres/src/main/scala/com/github/imliar/slick/repository/postgresql/TableWithId.scala
@@ -1,7 +1,0 @@
-package com.github.imliar.slick.repository.postgresql
-
-import PostgresDriver.api._
-
-abstract class TableWithId[Id, R](tag: Tag, tableName: String) extends Table[R](tag, tableName) {
-  def id: Rep[Id]
-}
