--- slick-repository-postgres/src/test/scala/com/github/imliar/slick/repository/postgresql/AwaitHelper.scala
@@ -1,14 +1,0 @@
-package com.github.imliar.slick.repository.postgresql
-
-import scala.concurrent.{Await, Future}
-import scala.concurrent.duration.Duration
-
-trait AwaitHelper {
-
-  implicit class AwaitFuture[A](f: Future[A]) {
-
-    def await(implicit timeout: Duration): A = Await.result[A](f, timeout)
-
-  }
-
-}
