--- slick-repository-postgres/src/test/scala/com/github/imliar/slick/repository/postgresql/models/User.scala
@@ -1,3 +1,0 @@
-package com.github.imliar.slick.repository.postgresql.models
-
-case class User(id: Option[Long], name: String)
