+++ slicker-postgres/src/test/resources/logback.xml
