+++ slicker-postgres/src/test/scala/io/slicker/postgres/models/User.scala
@@ -1,0 +1,3 @@
+package io.slicker.postgres.models
+
+case class User(id: Option[Long], name: String)
